requirements.txt
aiohttp
async_timeout
uvloop
bs4
sanic >= 0.5.1
jinja2
lxml
motor
wechatpy
cryptography
gunicorn
arrow
pillow
requests
